+++ examples/morpheusvm/controller/controller.go
@@ -15,7 +15,6 @@ import (
 
 	"github.com/ava-labs/hypersdk/auth"
 	"github.com/ava-labs/hypersdk/chain"
-	"github.com/ava-labs/hypersdk/examples/morpheusvm/config"
 	"github.com/ava-labs/hypersdk/examples/morpheusvm/consts"
 	"github.com/ava-labs/hypersdk/examples/morpheusvm/genesis"
 	"github.com/ava-labs/hypersdk/examples/morpheusvm/rpc"
@@ -79,7 +78,7 @@ func (*factory) New(
 	}
 
 	// Load config and genesis
-	c.config, err = config.New(configBytes)
+	c.config, err = newConfig(configBytes)
 	if err != nil {
 		return nil, nil, nil, err
 	}
@@ -134,7 +133,7 @@ type Controller struct {
 	chainID   ids.ID
 
 	genesis      *genesis.Genesis
-	config       *config.Config
+	config       *Config
 	stateManager *storage.StateManager
 
 	metrics *metrics
